JDBC - Drop custom class support until we can resolve the class loader issues
Due to class loading issues, setting a custom 'class' on a datasource will just not work. We can take another pass at this, but for now, "removing" support fixes existing datasource configs with a 'class' property already specified. We already use the driver and JDBC url to connect, and setting a custom class will only lock you to a specific DB/vendor version.