JDBC - Tweak generated key handling to avoid erroring when pulling generated keys
MSSQL loves to throw errors if a generated key doesn't exist. (Instead of simply returning null as other DB vendors do.) Since MSSQL also supports multiple languages, it's not easily possible to detect the reason for the error. All in all, we should log a warning and move on since the query DID complete successfully.

Resolves BL-884.